+++ tools/SetupFlow/DevHome.SetupFlow/TaskOperator/InstallOperator.cs
@@ -92,18 +92,3 @@ public class InstallOperator : IInstallOperator
         }).AsAsyncOperation();
     }
 }
-
-public class InstallPackageResult : IInstallPackageResult
-{
-    public int ExtendedErrorCode { get; set; }
-
-    public uint InstallerErrorCode { get; set; }
-
-    public int Status { get; set; }
-
-    public bool Attempted { get; set; }
-
-    public bool RebootRequired { get; set; }
-
-    public bool Succeeded { get; set; }
-}
